【开发】代码优化

+++ src/Endian.h
@@ -61,13 +61,13 @@ inline uint64_t networkToHost64(uint64_t net64)
     return be64toh(net64);
 }
 
-inline uint64_t networkToHost32(uint32_t net64)
+inline uint32_t networkToHost32(uint32_t net64)
 {
     return be32toh(net64);
 }
 
 // 从big-endian顺序转换为主机字节顺序
-inline uint64_t networkToHost16(uint16_t net16)
+inline uint16_t networkToHost16(uint16_t net16)
 {
     return be16toh(net16);
 }
